requirements.txt
rich==13.8.0
python-dotenv==1.0.1
langchain==0.3.7
langchain-openai==0.2.5
